+++ test/Main.hs
@@ -1,5 +1,8 @@
 module Main (main) where
 
+import Kafka.Effectful.Consumer.ClassifyTest qualified as ClassifyTest
+import Kafka.Effectful.Consumer.InterpreterTest qualified as InterpreterTest
+import Kafka.Effectful.OpenTelemetry.ConsumerSpanTest qualified as ConsumerSpanTest
 import Kafka.Effectful.OpenTelemetry.PropagationTest qualified as PropagationTest
 import Kafka.Effectful.OpenTelemetry.SemanticTest qualified as SemanticTest
 import Kafka.Effectful.OpenTelemetry.ShibuyaCompatibilityTest qualified as ShibuyaCompatibilityTest
@@ -11,8 +14,17 @@ main = defaultMain tests
 tests :: TestTree
 tests =
     testGroup
-        "Kafka.Effectful.OpenTelemetry"
-        [ SemanticTest.tests
-        , PropagationTest.tests
-        , ShibuyaCompatibilityTest.tests
+        "kafka-effectful"
+        [ testGroup
+            "Kafka.Effectful.OpenTelemetry"
+            [ SemanticTest.tests
+            , PropagationTest.tests
+            , ShibuyaCompatibilityTest.tests
+            , ConsumerSpanTest.tests
+            ]
+        , testGroup
+            "Kafka.Effectful.Consumer"
+            [ ClassifyTest.tests
+            , InterpreterTest.tests
+            ]
         ]